Increase nofile limit.
This will default to 16384, 16x what it was previously. This could go higher, but that seems like a reasonable limit for a modern machine.

Added a test for the nofile limit change. Fixed an existing test to account for the binary path change.

Added a directive to create the `/var/log/tapalcatl` directory, since converging a new instance seemed to fail when that was missing.